+++ src/lib.js
@@ -1,7 +1,8 @@
 
 export { ul_list, ol_list } from './list_func.js';
-export { link } from "./link_func.js";
+export { router_link } from "./func_router_link.js";
 export { image } from "./image_func.js";
 export { static_table, static_row, static_cell } from "./static_table_func.js";
 export { span } from "./span_func.js";
 export { div } from "./div_func.js";
+export { link } from "./func_link.js";